+++ tools/addstring.sh
@@ -1,4 +1,3 @@
 #!/bin/bash
-grep -Rl "\<string name=\"help_dangerous" . | xargs sed -i -e '/string name="msg_filtering/d
-'
-
+grep -RIl "\<string name=\"msg_usage" . | xargs sed -i -e '/string name="msg_usage/a \
+\ \ \ \ <string name="msg_restart">Application restart required</string>'